requirements.txt
pillow
opencv-python
numpy
imutils
matplotlib